+++ inst/python/rpytools/thread.py
@@ -1,28 +1,19 @@
-import rpycall
-import threading
-
 import sys
-
-is_py2 = sys.version[0] == "2"
-if is_py2:
-    import Queue as queue
-else:
-    import queue as queue
+import queue as queue
+from rpycall import call_r_function, schedule_python_function_on_main_thread
 
 
-def main_thread_func(f):
+def safe_call_r_function(f, args, kwargs):
+    try:
+        return call_r_function(f, *args, **kwargs), None
+    except Exception as e:  # TODO: should we catch BaseException too? KeyboardInterrupt?
+        return None, e
 
-    def python_function(*args, **kwargs):
 
-        if isinstance(threading.current_thread(), threading._MainThread):
-            res = f(*args, **kwargs)
-        else:
-            result = queue.Queue()
-            rpycall.call_python_function_on_main_thread(
-                lambda: result.put(f(*args, **kwargs)), None
-            )
-            res = result.get()
-
-        return res
-
-    return python_function
+def safe_call_r_function_on_main_thread(f, *args, **kwargs):
+    result = queue.Queue()
+    schedule_python_function_on_main_thread(
+        lambda: result.put(safe_call_r_function(f, args, kwargs)),
+        None,
+    )
+    return result.get()