+++ tests/testthat/test-python-threads.R
@@ -38,3 +38,53 @@ def write_to_file_from_thread(path, lines):
   expect_identical(readLines(file), letters)
 
 })
+
+
+
+test_that("Python calls into R from a background thread are evaluated", {
+
+  x <- 0L
+  py$r_func <- function() x <<- x+1
+  on.exit(py_del_attr(py, "r_func"), add = TRUE)
+  py_file <- withr::local_tempfile(lines = "r_func()", fileext = ".py")
+
+  reticulate:::py_run_file_on_thread(py_file)
+
+  # Simulate the main R thread doing non-Python work (e.g., sleeping)
+  for(i in 1:10) {
+    Sys.sleep(.01 * i)
+    if (x != 0L) break
+  }
+
+  expect_equal(x, 1L)
+})
+
+
+test_that("Errors from background threads calling into main thread are handled", {
+
+  py$signal_r_error <- function() stop("foo-bar-baz")
+  on.exit(py_del_attr(py, "signal_r_error"), add = TRUE)
+
+  # when testing, `r` in Python resolves `getOption("rlang_trace_top_env"), not
+  # the R globalenv(). Avoid using it to communicate.
+  val <- NULL
+  py$set_val <-  function(v) val <<- v
+  on.exit(py_del_attr(py, "set_val"), add = TRUE)
+
+  py_file <- withr::local_tempfile(lines = "
+try: signal_r_error()
+except Exception as e: set_val(e.args[0])
+", fileext = ".py")
+
+  reticulate:::py_run_file_on_thread(py_file)
+
+  # Simulate the main R thread doing non-Python work (e.g., sleeping)
+  for(i in 1:10) {
+    Sys.sleep(.01 * i)
+    if(!is.null(val)) break
+  }
+
+  expect(!is.null(val), "`thread_err_msg` never created")
+  expect_equal(val, "foo-bar-baz")
+
+})